+++ src/lightning/app/cli/commands/app_commands.py
@@ -1,4 +1,4 @@
-# Copyright The Lightning team.
+# Copyright The Lightning AI team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@ from typing import Dict, Optional
 
 import requests
 
-from lightning.app.cli.commands.connection import (
+from lightning.app.cli.connect.app import (
     _clean_lightning_connection,
     _install_missing_requirements,
     _resolve_command_path,