+++ src/lightning/pytorch/trainer/connectors/logger_connector/fx_validator.py
@@ -1,4 +1,4 @@
-# Copyright The Lightning team.
+# Copyright The Lightning AI team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -142,15 +142,6 @@ class _FxValidator:
         ),
         "test_step_end": _LogOptions(
             allowed_on_step=(False, True), allowed_on_epoch=(False, True), default_on_step=False, default_on_epoch=True
-        ),
-        "training_epoch_end": _LogOptions(
-            allowed_on_step=(False,), allowed_on_epoch=(True,), default_on_step=False, default_on_epoch=True
-        ),
-        "validation_epoch_end": _LogOptions(
-            allowed_on_step=(False,), allowed_on_epoch=(True,), default_on_step=False, default_on_epoch=True
-        ),
-        "test_epoch_end": _LogOptions(
-            allowed_on_step=(False,), allowed_on_epoch=(True,), default_on_step=False, default_on_epoch=True
         ),
         "configure_optimizers": None,
         "train_dataloader": None,