+++ tests/tests_pytorch/loggers/test_utilities.py
@@ -1,4 +1,4 @@
-# Copyright The PyTorch Lightning team.
+# Copyright The Lightning AI team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from lightning.pytorch.loggers import CSVLogger
-from lightning.pytorch.utilities.logger import _version
+from lightning.pytorch.loggers.utilities import _version
 
 
 def test_version(tmpdir):